refactor: move into util folder
